+++ package/contents/ui/config/configSearch.qml
@@ -148,7 +148,7 @@ KCM.ScrollViewKCM {
     }
 
     function setHeaders(xhr) {
-        xhr.setRequestHeader("User-Agent", "OnAir/2026.37")
+        xhr.setRequestHeader("User-Agent", "OnAir/2026.38")
     }
 
     function getStations(by, val) {